Perms-and-roles

+++ app/Http/Controllers/ProductController.php
@@ -9,6 +9,7 @@ use App\Models\Category;
 
 class ProductController extends Controller
 {
+    
     /**
      * Display a listing of the resource.
      *
@@ -47,14 +48,14 @@ class ProductController extends Controller
             'name' => $request->name,
             'description' => $request->description,
             'price' => $request->price,
-        ]);
+        ])->user()->associate($request->user());
+
+        $product->save();
 
         return response()->json([
             "status" => "success",
             "message" => "Product created successfully",
             "Product" => $product], 201);
-
-        
     }
 
     /**
@@ -113,10 +114,17 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        $product->delete();
+        if(auth()->user()->can('update all products') || (auth()->user()->can('update own products') && $product->user()->is(auth()->user()))) {
+            
+            $product->delete();
+    
+            return response()->json([
+                "status" => "success",
+                "message" => "Product deleted successfully"], 200);
+            }
+            else return response()->json([
+                "status" => "error",
+                "message" => "You are not authorized to delete this product"], 403);
 
-        return response()->json([
-            "status" => "success",
-            "message" => "Product deleted successfully"], 200);
     }
 }
